+++ atomicshop/wrappers/process_wrapper_pbtk.py
@@ -1,6 +1,3 @@
-# v1.0.2 - 21.03.2023 13:30
-import shlex
-
 from atomicshop.process import execute_with_live_output
 
 
@@ -14,7 +11,6 @@ def execute_pbtk(pbtk_path: str, file_path: str, target_directory: str) -> None:
     """
 
     cmd: str = f'python "{pbtk_path}" "{file_path}" "{target_directory}"'
-    cmd_list: list = shlex.split(cmd)
     print(f'Scanning: {file_path}')
 
-    output_list = execute_with_live_output(cmd=cmd_list)
+    output_list = execute_with_live_output(cmd=cmd)